Added stricter build config file, with settings based on http://rentzsch.tumblr.com/post/237349423/hoseyifyxcodewarnings-scpt (Thanks to lhunath for the idea - http://github.com/lhunath/asi-http-request/commit/d28fe4062bc2bb665e3d5828e2266584800f30d2)
Made some changes to fix warnings
Have disabled some settings, may re-enable in future:
* Need formal protocols for delegates, as per llunath's fork
* Should consider renaming S3 methods to prevent the clang warnings

Fixed another issue with test breakage on Mac

+++ Classes/ASIFormDataRequest.m
@@ -181,7 +181,7 @@
 	NSString *endItemBoundary = [NSString stringWithFormat:@"\r\n--%@\r\n",stringBoundary];
 	NSEnumerator *e = [[self postData] keyEnumerator];
 	NSString *key;
-	int i=0;
+	NSUInteger i=0;
 	while (key = [e nextObject]) {
 		[self appendPostString:[NSString stringWithFormat:@"Content-Disposition: form-data; name=\"%@\"\r\n\r\n",key]];
 		[self appendPostString:[[self postData] objectForKey:key]];
